Enum rutabaga_gfx::RutabagaError

pub enum RutabagaError {
    AlreadyInUse,
    BaseError(BaseError),
    CheckedArithmetic {
        field1: (&'static str, usize),
        field2: (&'static str, usize),
        op: &'static str,
    },
    CheckedRange {
        field1: (&'static str, usize),
        field2: (&'static str, usize),
    },
    ComponentError(i32),
    Invalid2DInfo,
    InvalidCapset,
    InvalidCommandSize(usize),
    InvalidComponent,
    InvalidContextId,
    InvalidCrossDomainChannel,
    InvalidCrossDomainItemId,
    InvalidCrossDomainItemType,
    InvalidCrossDomainState,
    InvalidGrallocBackend,
    InvalidGrallocDimensions,
    InvalidGrallocDrmFormat,
    InvalidGrallocGpuType,
    InvalidGrallocNumberOfPlanes,
    InvalidIovec,
    InvalidResourceId,
    InvalidRutabagaBuild,
    InvalidRutabagaHandle,
    InvalidVulkanInfo,
    IoError(IoError),
    MappingFailed(ExternalMappingError),
    SpecViolation(&'static str),
    TryFromIntError(TryFromIntError),
    Unsupported,
    Utf8Error(Utf8Error),
    VolatileMemoryError(VolatileMemoryError),
}

An error generated while using this crate.

Variants

AlreadyInUse

Indicates Rutabaga was already initialized since only one Rutabaga instance per process is allowed.

BaseError(BaseError)

Base error returned as a result of rutabaga library operation.

CheckedArithmetic

Checked Arithmetic error

Fields of CheckedArithmetic

field1: (&'static str, usize)field2: (&'static str, usize)op: &'static str

CheckedRange

Checked Range error

Fields of CheckedRange

field1: (&'static str, usize)field2: (&'static str, usize)

ComponentError(i32)

An internal Rutabaga component error was returned.

Invalid2DInfo

Invalid 2D info

InvalidCapset

Invalid Capset

InvalidCommandSize(usize)

A command size was submitted that was invalid.

InvalidComponent

Invalid RutabagaComponent

InvalidContextId

Invalid Context ID

InvalidCrossDomainChannel

Invalid cross domain channel

InvalidCrossDomainItemId

Invalid cross domain item ID

InvalidCrossDomainItemType

Invalid cross domain item type

InvalidCrossDomainState

Invalid cross domain state

InvalidGrallocBackend

Invalid gralloc backend.

InvalidGrallocDimensions

Invalid gralloc dimensions.

InvalidGrallocDrmFormat

Invalid gralloc DRM format.

InvalidGrallocGpuType

Invalid GPU type.

InvalidGrallocNumberOfPlanes

Invalid number of YUV planes.

InvalidIovec

The indicated region of guest memory is invalid.

InvalidResourceId

Invalid Resource ID.

InvalidRutabagaBuild

Indicates an error in the RutabagaBuilder.

InvalidRutabagaHandle

An error with the RutabagaHandle

InvalidVulkanInfo

Invalid Vulkan info

IoError(IoError)

An input/output error occured.

MappingFailed(ExternalMappingError)

The mapping failed.

SpecViolation(&'static str)

Violation of the Rutabaga spec occured.

TryFromIntError(TryFromIntError)

An attempted integer conversion failed.

Unsupported

The command is unsupported.

Utf8Error(Utf8Error)

Utf8 error.

VolatileMemoryError(VolatileMemoryError)

Volatile memory error