Module rutabaga_gfx::rutabaga_core

rutabaga_core: Cross-platform, Rust-based, Wayland and Vulkan centric GPU virtualization.

Structs

Rutabaga

The global libary handle used to query capability sets, create resources and contexts.

Rutabaga2DInfo

Information required for 2D functionality.

RutabagaBuilder

Rutabaga Builder, following the Rust builder pattern.

RutabagaCapsetInfo

RutabagaResource

A Rutabaga resource, supporting 2D and 3D rutabaga features. Assumes a single-threaded library.

Constants

RUTABAGA_CAPSETS

Traits

RutabagaComponent

A RutabagaComponent is a building block of the Virtual Graphics Interface (VGI). Each component on it’s own is sufficient to virtualize graphics on many Google products. These components wrap libraries like gfxstream or virglrenderer, and Rutabaga’s own 2D and cross-domain prototype functionality.

RutabagaContext

Functions

calculate_context_mask