Struct rutabaga_gfx::rutabaga_core::RutabagaBuilder

pub struct RutabagaBuilder {
    display_width: Option<u32>,
    display_height: Option<u32>,
    default_component: RutabagaComponentType,
    gfxstream_flags: GfxstreamFlags,
    virglrenderer_flags: VirglRendererFlags,
    context_mask: u64,
    channels: Option<Vec<RutabagaChannel>>,
}

Rutabaga Builder, following the Rust builder pattern.

Fields
display_width: Option<u32>display_height: Option<u32>default_component: RutabagaComponentTypegfxstream_flags: GfxstreamFlagsvirglrenderer_flags: VirglRendererFlagscontext_mask: u64channels: Option<Vec<RutabagaChannel>>
Implementations

impl RutabagaBuilder

pub fn new(
default_component: RutabagaComponentType,
context_mask: u64
) -> RutabagaBuilder

Create new a RutabagaBuilder.

pub fn set_display_width(self, display_width: u32) -> RutabagaBuilder

Set display width for the RutabagaBuilder

pub fn set_display_height(self, display_height: u32) -> RutabagaBuilder

Set display height for the RutabagaBuilder

pub fn set_use_egl(self, v: bool) -> RutabagaBuilder

Sets use EGL flags in gfxstream + virglrenderer.

pub fn set_use_gles(self, v: bool) -> RutabagaBuilder

Sets use GLES in gfxstream + virglrenderer.

pub fn set_use_glx(self, v: bool) -> RutabagaBuilder

Sets use GLX flags in gfxstream + virglrenderer.

pub fn set_use_surfaceless(self, v: bool) -> RutabagaBuilder

Sets use surfaceless flags in gfxstream + virglrenderer.

pub fn set_use_vulkan(self, v: bool) -> RutabagaBuilder

Sets use Vulkan in gfxstream + virglrenderer.

pub fn set_use_guest_angle(self, v: bool) -> RutabagaBuilder

Set use guest ANGLE in gfxstream

pub fn set_use_external_blob(self, v: bool) -> RutabagaBuilder

Sets use external blob in virglrenderer.

pub fn set_use_render_server(self, v: bool) -> RutabagaBuilder

Sets use render server in virglrenderer.

pub fn set_wsi(self, v: Option<&RutabagaWsi>) -> RutabagaBuilder

Use the Vulkan swapchain to draw on the host window for gfxstream.

pub fn set_rutabaga_channels(
self,
channels: Option<Vec<RutabagaChannel>>
) -> RutabagaBuilder

Set rutabaga channels for the RutabagaBuilder

pub fn build(
self,
fence_handler: RutabagaFenceHandler,
render_server_fd: Option<SafeDescriptor>
) -> RutabagaResult<Rutabaga>

Builds Rutabaga and returns a handle to it.

This should be only called once per every virtual machine instance. Rutabaga tries to intialize all 3D components which have been built. In 2D mode, only the 2D component is initialized.