Module rutabaga_gfx::virgl_renderer

virgl_renderer: Handles 3D virtio-gpu hypercalls using virglrenderer. External code found at https://gitlab.freedesktop.org/virgl/virglrenderer/.

Structs

VirglRenderer

The virtio-gpu backend state tracker which supports accelerated rendering.

VirglRendererContext

Constants

VIRGL_RENDERER_CALLBACKS

Functions

debug_callback

export_query

Retrieves metadata suitable for export about this resource. If “export_fd” is true, performs an export of this resource so that it may be imported by other processes.

import_resource

map_func

unmap_func

write_context_fence

TODO(ryanneph): re-evaluate if “ring_idx: u8” can be used instead so we can drop this in favor of the common write_context_fence() from renderer_utils before promoting to cfg(feature = “virgl_renderer”).

Type Definitions

Query