Trait rutabaga_gfx::rutabaga_core::RutabagaContext

pub trait RutabagaContext {
    fn submit_cmd(&mut self, _commands: &mut [u8]) -> RutabagaResult<()>;

    fn attach(&mut self, _resource: &mut RutabagaResource);

    fn detach(&mut self, _resource: &RutabagaResource);

    fn component_type(&self) -> RutabagaComponentType;

    fn context_create_blob(
        &mut self, 
        _resource_id: u32, 
        _resource_create_blob: ResourceCreateBlob, 
        _handle_opt: Option<RutabagaHandle>
    ) -> RutabagaResult<RutabagaResource> { ... }

    fn context_create_fence(
        &mut self, 
        _fence: RutabagaFence
    ) -> RutabagaResult<()> { ... }

    fn context_poll(&mut self) -> Option<Vec<RutabagaFence>> { ... }
}

Required methods

fn submit_cmd(&mut self, _commands: &mut [u8]) -> RutabagaResult<()>

Implementations must handle the context-specific command stream.

fn attach(&mut self, _resource: &mut RutabagaResource)

Implementations may use resource in this context’s command stream.

fn detach(&mut self, _resource: &RutabagaResource)

Implementations must stop using resource in this context’s command stream.

fn component_type(&self) -> RutabagaComponentType

Implementations must return the component type associated with the context.

Provided methods

fn context_create_blob(
&mut self,
_resource_id: u32,
_resource_create_blob: ResourceCreateBlob,
_handle_opt: Option<RutabagaHandle>
) -> RutabagaResult<RutabagaResource>

Implementations must return a RutabagaResource given the resource_create_blob parameters.

fn context_create_fence(&mut self, _fence: RutabagaFence) -> RutabagaResult<()>

Implementations must create a fence on specified ring_idx in fence. This allows for multiple synchronizations timelines per RutabagaContext.

fn context_poll(&mut self) -> Option<Vec<RutabagaFence>>

Implementations must return an array of fences that have completed. This will be used by the cross-domain context for asynchronous Tx/Rx.