Module rutabaga_gfx::rutabaga_gralloc::gralloc

gralloc: Cross-platform, Rust-based, Vulkan centric GPU allocation and mapping.

Structs

ImageAllocationInfo

Information required to allocate a swapchain image.

ImageMemoryRequirements

The memory requirements, compression and layout of a swapchain image.

RutabagaGralloc

A container for a variety of allocation backends.

RutabagaGrallocFlags

Usage flags for constructing a buffer object.

Enums

GrallocBackend

Enumeration of possible allocation backends.

Constants

RUTABAGA_GRALLOC_USE_CAMERA_READ

RUTABAGA_GRALLOC_USE_CAMERA_WRITE

RUTABAGA_GRALLOC_USE_LINEAR

RUTABAGA_GRALLOC_USE_PROTECTED

RUTABAGA_GRALLOC_USE_RENDERING

RUTABAGA_GRALLOC_USE_SCANOUT

RUTABAGA_GRALLOC_USE_SW_READ_OFTEN

RUTABAGA_GRALLOC_USE_SW_WRITE_OFTEN

RUTABAGA_GRALLOC_USE_TEXTURING

RUTABAGA_GRALLOC_VIDEO_DECODER

RUTABAGA_GRALLOC_VIDEO_ENCODER

Traits

Gralloc

Trait that needs to be implemented to service graphics memory requests. Two step allocation process: